attempts to port setup script to use numpy.distutils.core setup have failed.
Upgrading the setup.py script to replace setuptools setup function
with the numpy.distutils.core setup have failed and thus have
left me with a bunch of commented out and otherwise stray and unused
code throughout the entire package tree as it calls for creating
a setup.py file in each subpackage. I will leave this for now in hopes
of someday resolving it but for now its back to using the
setuptools setup function.