changing __init__ method of Atom class.
Updating to make sure that symbol and atomic number are assigned
a value other than None. Replaced the try...except with simpler
if...else as the logic for possible scenarios is not complex and much
more readable in this form.

Need to write some more tests and benchmark code to be sure this is
better.