Adding call to clear() method on `structure_atoms` attribute.
Calling the read() method appends atoms to the `structure_atoms`
list. Thus, calling read() more than once effectively
extends the list of atoms with a duplicate instead of updating
the list. Therefore, adding the call to clear() to the read()
method to reset the list of atoms to an empty list.