Adding new attributes `PG` for the Point Group and `Ni`.
Some fullerene isomers are identical in `N` and point group, but
their symmetry is distinguished by some other symmetry element,
with file names distinguished with a "-Ni" number, where
"Ni" is an integer >= 0. Therefore, if an `Ni` is given
to the constructor call, it will be used to filter the
list of fullerene data files to choose from.