Removing automatic alls to `update_methods`.
Prior to this commit, the property decorator methods
called an `update_*` method to update atom attributes
based on an atom_tree query before returning the property
list of all atoms in the list. This leads to many unnecessary
repeated calls to the same update method.

This commit renames all former public update methods to
private methods (`_update_*`) and removes the automatic
calls to those methods (for now, by commenting the line out)
and adds a new public update method `update_attrs` which
calls all `_update_*` methods necessary to update atom
attributes, but only once.

It's now up to the user to call this method to update
the atom attributes.