Adding a class variable to hold `Atom` and `Atoms` attr strings.
Using class variable `attributes` list to hold the attribute
string names of the class. Each subclass extends the parent
class's `attributes` list with its respective attributes names.
The list is for maintaining the list of attributes that should
be recognized by the method `getatomattr` which iterates through
the list of `Atom` objects, gets the `Atom` attribute if its
in the `attributes` list, and returns the populated list
as an array.