Suspending further development on analyze_structure script.
The analysis code was becoming too customized and tailored to
a specific application, which is inconsistent with the
intent of the scikit-nano scripts. The scripts should provide
a command-line interface for executing the functions and classes
that are utilized in an interactive session, or to run any
GUI apps, which is what the `nanogen` and `nanogenui` scripts
do, respectively.

If I further develop the `StructureAnalyzer` class with general
structure analysis methods, then this script should be developed
further as a CLI to that class.