Refactoring crystallography base classes.
At the cost of redundant calculations, I decided to add
"symmetric" class definitions for all direct and reciprocal
lattice objects. Therefore, instead of just a
`ReciprocalLatticeMixin` class which `CrystalLattice` subclassed,
there is now a `ReciprocalLattice` class and a
`DirectLatticeMixin` class which is subclassed by the
`ReciprocalLattice` class. I don't know, maybe this was
dumb, but for now my OCD is demanding symmetry in the
direct/reciprocal lattice classes.