Changing the default value for `KDTAtoms.kNN` to 16.
Formerly the default was set to 3, which was due to
my initial interest in dealing with pristine carbon
nanotubes and graphene. However, this was preventing an
accurate count of nearest neighbors statistics when
looking at bulk structures or nanostructures with
adatoms, etc. The new default should address this,
but perhaps it should be a positional parameter to
remind the user to set this to a value needed for
analysis?