Updating the `Bond` class.
Adding a new lazily computed `centroid` property to
return the bond centroid from the 2 atoms positions.

Adding a new `rotate` method to rotate the bond by rotating
the bond atoms.

Updating the `Bond` class constructor to
create an empty `StructureAtoms` instance before extending
the list with the 2 atoms in the bond so as to
retain references to the original atom instances.

Also removing explicit setting of the `Bond.vector`
attribute in the constructor and instead making it a
lazily computed property.