Updating `UserList` class to sub-class `BaseClass`.
`UserList` now serves as an abstract base class with
a new abstract property `__item_class__` as well as the
inherited abstract method `todict` from `BaseClass`.

Updating the `__get_item__` and `__set_item__` methods
to use the `__item_class__` to attempt to cast the item
values to the appropriate item class defined by the
implementing class.

These changes are were made in an effort to cut down on
redundant code in `UserList` sub-classes.